Large volumes generate more than 32k waits, so need a full int for waits

+++ src/task.h
@@ -132,7 +132,7 @@ struct task {
   short int nr_unlock_tasks;
 
   /*! Number of unsatisfied dependencies */
-  short int wait;
+  int wait;
 
   /*! Type of the task */
   enum task_types type;
